AD_Drone/requirements.txt
kafka-python
python-dotenv
requests
cryptography
pycryptodome